+++ src/middlewared/middlewared/plugins/system_advanced/syslog.py
@@ -9,20 +9,6 @@ class SystemAdvancedService(Service):
         cli_namespace = 'system.advanced'
 
     @accepts()
-    @returns(Dict('serial_port_choices', additional_attrs=True))
-    async def serial_port_choices(self):
-        """
-        Get available choices for `serialport`.
-        """
-        ports = {e['name']: e['name'] for e in await self.middleware.call('device.get_info', 'SERIAL')}
-        if not ports or (await self.middleware.call('system.advanced.config'))['serialport'] == 'ttyS0':
-            # We should always add ttyS0 if ports is false or current value is the default one in db
-            # i.e ttyS0
-            ports['ttyS0'] = 'ttyS0'
-
-        return ports
-
-    @accepts()
     @returns(Dict(
         additional_attrs=True,
         title='Syslog Certificate Choices',
